Move iptvcat files into separate folder
